Rename test suite to fit griddaemons login coverage.
